Move job selector test to separate file
Remove need for cleanup
